kernel/makefile: Avoid repeating download of source tar on rebuilds
This fixes kernel/makefile to skip downloading the source tar
and/or extracting it during kernel rebuilds, it its not needed
(i.e. when $(KERNEL_EXTRACTED) already exists).

The fix consists in turning the directory prerequisites
from "normal" to "order-only" prerequisites, so that any timestamp
change in the required folders doesn't uselessly trigger
the make'ing of $(KERNEL_EXTRACTED) target.

Also, this commit adds $(KERNEL_TAR_SIGNATURE) target
as a prerequisite of $(KERNEL_EXTRACTED) to ensure that the
corresponding signature file always exists for a given source tar.
This fixes the "file not found" error when gpg-verifying
the downloaded tar when building a new kernel version.

Signed-off-by: Fil <fil@filberg.eu>