LibKernelUpgrade.sh: Fix invocation of get_build_kver
This fixes both the behaviour of get_build_kver() and its invocation in other
parts of the script, so that it executes in the same shell as the caller's.
This way, all parts of the version string can be retrieved in a single call.
Before this fix, the function had to be invoked in a subshell $(foo)
to capture the full version string as its output, while the export'ed values
of $PRAWNOS_KVER_* could not be read by the caller in the parent shell.

Signed-off-by: Fil <fil@filberg.eu>