add support for rk3588 server board with uboot bootloader
TARGET is arm64-rk3588-server
The server images are more lightweight than the normal images.
To support root disk encryption, the server initramfs contains an
openssh server to remotely decrypt the root drive on boot.

Other changes:

- Fixup clean for initramfs/kernel/kernel package

Previously, we would try to keep around the kernel orig.tar.gz
to make package reuploads easier
But that overcomplicates cleaning the kernel package, and is generally a
bad practice.
Stop doing that, and fix up the related clean actions

- add ripgrep to all images
- add htop to all images
- remove many unused arm64 kernel patches

History for this commit can be found on the PrawnOS/rk3588 branch